Fixing setup logic.
Moving all project metadata (version, author, etc) into
pycoevolity/metadata.py. This is then read within setup.py to get the
metadata for the setup function, and is imported by
pycoevolity/__init__.py, so that it is available within the package.

This was needed because I was doing 'from pycoevolity import
__version__' from within setup.py. This broke once I added a dependency
(munkres), because pycoevolity could not be imported without it, and
munkres was not installed yet (the import was at the top of setup.py.).