Adding option for label changes to loci2alignment
Updating CLI tools based on loci2alignment.py allowing a CSV file to be
provided for changing the seq labels.

This required updates to the Loci class in parsing to parse the CSV
file, store the name changes in a dict, and convert the names will
parsing the data files.

All tests are passing.