Updating CLI concatenation tools.
Adding options to fastas2alignment.py and loci2alignment.py.

Adding a 'subsample' option to specify a number of loci to be
randomly subsampled without replacement.

Also, adding the option to specify a tabular file with name changes for
fastas2alignment.py (this option was already in loci2alignment.py).